refactor(server): response package
- move response.go in package response
- split reponse.go in files by purpose
